+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## 2.4.0 - November 2023
 
 1. fix: project link font weight change on hover
+2. feat!: gradient background for contours and remove source link
 
 ## 2.3.0 - November 2023